Refactor project model and view directory
